W111MdlStub: reconstruct six mdlFile functions (obj 8->248/368 nd 8->197; 56->96/288 nd31->46; 88->364/368 nd52->304; 152->292/304 nd120->228; 292->584/592 nd215->399; 484->1044/1056 nd343->727)

+++ src/Event/Fcl/fclShopMisc.c
@@ -268,7 +268,7 @@ void FUN_003f7d50(int param_1,int param_2,u64 param_3,int param_4, int param_5);
 void FUN_003f7fe0(u64 param_1,u64 param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f8180(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f85a0(int param_1,int param_2,u64 param_3,int param_4, int param_5);
-void FUN_003f86a0(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
+void FUN_003f86a0(int param_1,int param_2,int param_3,int param_4,int param_5 );
 void FUN_003f8a00(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
 void FUN_003f8e10(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
 void FUN_003f9220(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
@@ -7282,7 +7282,7 @@ void FUN_003f85a0(int param_1,int param_2,u64 param_3,int param_4,
 // FUN_003F86A0 NONMATCHING
 
 
-void FUN_003f86a0(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5
+void FUN_003f86a0(int param_1,int param_2,int param_3,int param_4,int param_5
 
                  )
